summarizer/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
youtube-transcript-api>=1.2.3
pytubefix>=1.6.3
groq>=0.4.2
openai>=1.3.7
aiohttp>=3.9.0
requests>=2.31.0
wget>=3.2
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.1
dropbox>=11.36.2
ffmpeg-python>=0.2.0
